+++ .env.example
@@ -8,6 +8,13 @@ APP_TITLE=LibreChat
 # Uncomment and make empty "" to remove the footer.
 # CUSTOM_FOOTER="My custom footer"
 
+# Uncomment the `ENDPOINTS` variable to determine which endpoints are available.
+# Listed here are the only accepted values, which are all enabled if an authentication
+# option is provided, whether user provided or admin provided in this .env file.
+# Note: the first value is considered the default value.
+
+# ENDPOINTS=openAI,azureOpenAI,bingAI,chatGPTBrowser,google,gptPlugins,anthropic
+
 # The server will listen to localhost:3080 by default. You can change the target IP as you want. 
 # If you want to make this server available externally, for example to share the server with others 
 # or expose this from a Docker container, set host to 0.0.0.0 or your external IP interface.
