+++ api/server/controllers/EndpointController.js
@@ -1,6 +1,6 @@
-const { EModelEndpoint } = require('../routes/endpoints/schemas');
-const { availableTools } = require('../../app/clients/tools');
-const { addOpenAPISpecs } = require('../../app/clients/tools/util/addOpenAPISpecs');
+const { EModelEndpoint } = require('~/server/routes/endpoints/schemas');
+const { availableTools } = require('~/app/clients/tools');
+const { addOpenAPISpecs } = require('~/app/clients/tools/util/addOpenAPISpecs');
 const {
   openAIApiKey,
   azureOpenAIApiKey,
@@ -13,13 +13,13 @@ const {
   bingAI,
   chatGPTBrowser,
   anthropic,
-} = require('../services/EndpointService').config;
+} = require('~/server/services/EndpointService').config;
 
 let i = 0;
 async function endpointController(req, res) {
   let key, palmUser;
   try {
-    key = require('../../data/auth.json');
+    key = require('~/data/auth.json');
   } catch (e) {
     if (i === 0) {
       i++;
@@ -54,18 +54,42 @@ async function endpointController(req, res) {
       }
       : false;
 
-  res.send(
-    JSON.stringify({
-      [EModelEndpoint.openAI]: openAI,
-      // [EModelEndpoint.assistant]: assistant,
-      [EModelEndpoint.azureOpenAI]: azureOpenAI,
-      [EModelEndpoint.google]: google,
-      [EModelEndpoint.bingAI]: bingAI,
-      [EModelEndpoint.chatGPTBrowser]: chatGPTBrowser,
-      [EModelEndpoint.gptPlugins]: gptPlugins,
-      [EModelEndpoint.anthropic]: anthropic,
-    }),
-  );
+  let enabledEndpoints = [
+    EModelEndpoint.openAI,
+    EModelEndpoint.azureOpenAI,
+    EModelEndpoint.google,
+    EModelEndpoint.bingAI,
+    EModelEndpoint.chatGPTBrowser,
+    EModelEndpoint.gptPlugins,
+    EModelEndpoint.anthropic,
+  ];
+
+  const endpointsEnv = process.env.ENDPOINTS || '';
+  if (endpointsEnv) {
+    enabledEndpoints = endpointsEnv
+      .split(',')
+      .filter((endpoint) => endpoint?.trim())
+      .map((endpoint) => endpoint.trim());
+  }
+
+  const endpointConfig = {
+    [EModelEndpoint.openAI]: openAI,
+    [EModelEndpoint.azureOpenAI]: azureOpenAI,
+    [EModelEndpoint.google]: google,
+    [EModelEndpoint.bingAI]: bingAI,
+    [EModelEndpoint.chatGPTBrowser]: chatGPTBrowser,
+    [EModelEndpoint.gptPlugins]: gptPlugins,
+    [EModelEndpoint.anthropic]: anthropic,
+  };
+
+  const orderedAndFilteredEndpoints = enabledEndpoints.reduce((config, key, index) => {
+    if (endpointConfig[key]) {
+      config[key] = { ...(endpointConfig[key] ?? {}), order: index };
+    }
+    return config;
+  }, {});
+
+  res.send(JSON.stringify(orderedAndFilteredEndpoints));
 }
 
 module.exports = endpointController;
