+++ packages/data-provider/src/types.ts
@@ -122,11 +122,12 @@ export type TConfig = {
   availableTools?: [];
   plugins?: Record<string, string>;
   azure?: boolean;
+  order: number;
 };
 
 export type TModelsConfig = Record<string, string[]>;
 
-export type TEndpointsConfig = Record<string, TConfig | null>;
+export type TEndpointsConfig = Record<string, TConfig>;
 
 export type TUpdateTokenCountResponse = {
   count: number;
